+++ libftprintf/ft_digit_num.c
@@ -4,13 +4,12 @@
 ** ft_digit_num: подсчет количества цифр в числе.
 */
 
-
 int		ft_digit_num(long int n)
 {
 	int tmp;
 	int tmp2;
 
-	tmp2 = n;
+	tmp2 = (int)n;
 	tmp = 1;
 	if (n < 0)
 		++tmp;
